Rename PairingsData.ts
